Created voting Helper and moved custom class to other classes
