+++ currencyii/src/main/res/navigation/nav_graph.xml
@@ -69,20 +69,10 @@
         android:name="nl.tudelft.trustchain.currencyii.ui.bitcoin.JoinDAOFragment"
         android:label="Join DAO"
         tools:layout="@layout/fragment_join_network">
-        <action
-            android:id="@+id/toVotesFragment"
-            app:destination="@id/votesFragment">
-            <argument
-                android:name="title"
-                app:argType="string" />
-            <argument
-                android:name="amount"
-                app:argType="string" />
-        </action>
     </fragment>
     <fragment
         android:id="@+id/votesFragment"
-        android:name="nl.tudelft.trustchain.common.ui.VotesFragment"
+        android:name="nl.tudelft.trustchain.currencyii.ui.bitcoin.VotesFragment"
         android:label="Votes" />
     <fragment
         android:id="@+id/createSWFragment"
